src/requirements.txt
python-dotenv
autopep8
requests